Keep track of the pending providers
This information augments what we get from SAML metadata and allows us
to reflect pending providers that are not yet in metadata for whatever
reason. The structure of this data is an array that is basically the
same as the safire.json metadata: key. Entries should be removed from
the pending file if/when they are added into metadata.