Update POPI-related info; may-vs-must
Update the consent section to reflect both the advice we received from
Cliffe Dekker Hofmeyr, and the substantive work being done by GÉANT,
REFEDS, JISC, and others around the European GDPR. This softens the use
of the term "consent" in light of other more preferable mechanisms for
lawful data processing (we've already done this on the notification page
itself, for the same reason).

Add a new "may-vs-must" section that references RFC 2119 as the original
source of the terminology. Also include the advice we received from
Cliffe Dekker Hofmeyr around section 10.2, and the fact that interpreted
under South African common law it is already sufficient.