Remove registrationInstant
The requirement to include registrationInstant was a SHOULD in the
eduGAIN SAML profile and has been removed from the new version. It adds
no value for us, and is often incorrect because we forget to update it.
Therefore since it is not necessary, we should remove it.

This is was a non-normative example, so removing it here has no effect
on the MRPS. The change has also been applied retrospectively in the
registry, so it has no impact on existing registered entities. Thus
it is merely cosmetic, and does not require we bump the version number
of the MRPS.